+++ packages/vscode-plugin/src/process.ts
@@ -74,7 +74,14 @@ export class Process extends EventEmitter {
 
     this.logger.info(`Server process started with PID ${this.#process.pid}`);
 
-    this.#process.on('exit', (code) => this.emit('exit', code));
+    this.#process.on('exit', (code) => {
+      if (code === 0) {
+        this.logger.info('Server process exited normally with code 0');
+      } else {
+        this.logger.error(`Server process exited with code ${code}`);
+      }
+    });
+
     return await this.getServerLocation();
   }
   private async getServerLocation(): Promise<ServerLocation> {