+++ packages/vscode-plugin/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "Stryker Mutation Testing" extension will be documented in this file.
 
+## 0.1.0
+
+- Fixed support for local workspace directory settings.
+- Added a reload workspace command to the command palette.
+
 ## 0.0.2
 
 - Fixed an issue where GIFs in the README were not displayed on the VS Code Marketplace page for the extension.